# use global rand functions