--- a/.github/workflows/test.yml
+++ b/.github/workflows/test.yml
@@ -10,7 +10,8 @@ jobs:
     strategy:
       matrix:
         go:
-          - "1"
+          - "stable"
+          - "1.22.0-rc.1"
           - "1.21"
           - "1.20"
           - "1.19"