From Go 1.20, `math/rand` uses thread-local source if possible.
It is faster than using the global source because it avoid locking cost.

go version: go version go1.22rc1 darwin/arm64

```
goos: darwin
goarch: arm64
pkg: github.com/shogo82148/go-retry
                   │   .old.txt    │              .new.txt               │
                   │    sec/op     │   sec/op     vs base                │
Continue-10            535.0n ± 0%   564.7n ± 0%   +5.55% (p=0.000 n=10)
ContinueSuccess-10     3.724n ± 0%   3.725n ± 0%        ~ (p=0.723 n=10)
Do-10                  2.864µ ± 0%   2.985µ ± 0%   +4.24% (p=0.000 n=10)
DoSuccess-10           5.555n ± 1%   5.532n ± 1%        ~ (p=0.101 n=10)
Do_Parallel-10       14192.0n ± 0%   434.2n ± 1%  -96.94% (p=0.000 n=10)
geomean                214.1n        108.6n       -49.29%

                   │   .old.txt   │              .new.txt               │
                   │     B/op     │    B/op     vs base                 │
Continue-10          0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
ContinueSuccess-10   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Do-10                16.00 ± 0%     16.00 ± 0%       ~ (p=1.000 n=10) ¹
DoSuccess-10         0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Do_Parallel-10       16.00 ± 0%     16.00 ± 0%       ~ (p=1.000 n=10) ¹
geomean                         ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                   │   .old.txt   │              .new.txt               │
                   │  allocs/op   │ allocs/op   vs base                 │
Continue-10          0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
ContinueSuccess-10   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Do-10                1.000 ± 0%     1.000 ± 0%       ~ (p=1.000 n=10) ¹
DoSuccess-10         0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Do_Parallel-10       1.000 ± 0%     1.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                         ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```